Chuyen cac ham vao dung file, tien hon cho debug sau nay.
