+++ LibManager/LibManager.vcxproj.filters
@@ -4,6 +4,7 @@
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Menu.cpp" />
     <ClCompile Include="authentication.cpp" />
+    <ClCompile Include="account.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="account.h" />
